+++ NEWS.md
@@ -57,6 +57,7 @@
   avoid reading and pre-processing of ignored files (@mwaldstein)
 * Allow for any number of `#` to start a comment. Useful in ESS (#299, @prosoitos)
 * New equals_na_linter() (#143, #326, @jabranham)
+* Fixed plain-code-block bug in Rmarkdown (#252, @russHyde)
 
 # lintr 1.0.2 #
 * Fix tests to work with upcoming testthat release.